updated questions router, completed validation for new question post route.
