+++ aurora-project-starter/views/question-ask.pug
@@ -1,16 +1,16 @@
 extends layout.pug
 
 block content
-  form(action="/question" method="post")
+  form(action="/questions" method="post")
     input(type="hidden" name="_csrf" value=csrfToken)
     label(for="title") Your Question:
     input(type="text" name="title" required)
     label(for="topicId") Topic:
-    select(name="topicId")
+    select(name="topicId" required)
       each topic in topics
         option(value= topic.id)= topic.name
     label(for="expertiseId") Expertise Level:
-    select(name="expertiseId")
+    select(name="expertiseId" required)
       each expertise in expertises
         option(value= expertise.id)= expertise.description
     div
